+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irConst.kt
@@ -5,7 +5,6 @@ import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.util.SYNTHETIC_OFFSET
 import org.jetbrains.kotlin.ir.interpreter.toIrConst as originalToIrConst
-import org.jetbrains.kotlin.ir.interpreter.toIrConstOrNull as originalToIrConstOrNull
 
 public fun IrExpression.toIrConst(
     irType: IrType = type,
@@ -24,7 +23,8 @@ public fun IrExpression?.toIrConstOrNull(
     endOffset: Int = SYNTHETIC_OFFSET,
 ): IrConst<*>? =
     if (this != null && irType != null) {
-        originalToIrConstOrNull(irType = irType, startOffset = startOffset, endOffset = endOffset)
+        runCatching { toIrConst(irType = irType, startOffset = startOffset, endOffset = endOffset) }
+            .getOrNull()
     } else {
         null
     }